src/final_assessment/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.61.0
anyio==4.10.0
async-timeout==4.0.3
attrs==25.3.0
authlib==1.6.1
backports-tarfile==1.2.0
beautifulsoup4==4.13.4
brotli==1.1.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cryptography==45.0.6
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
docx2txt==0.9
duckduckgo-search==8.1.1
et-xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.116.1
ffmpy==0.6.1
filelock==3.18.0
filetype==1.2.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.29.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gradio==5.41.1
gradio-client==1.11.0
greenlet==3.2.4
groovy==0.1.2
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.3
idna==3.10
itsdangerous==2.2.0
jaraco-context==6.0.1
jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-experimental==0.3.4
langchain-google-genai==2.0.10
langchain-text-splitters==0.3.9
langgraph==0.6.3
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.3
langgraph-sdk==0.2.0
langsmith==0.4.13
llvmlite==0.44.0
lxml==6.0.0
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.5
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.6.3
mypy-extensions==1.1.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai==1.99.1
openai-whisper==20250625
openpyxl==3.1.5
orjson==3.11.1
ormsgpack==1.10.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
primp==0.15.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pydub==0.25.1
pygments==2.19.2
pyparsing==3.2.3
pypdf==5.9.0
pypdf2==3.0.1
pysocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-magic==0.4.27
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
rsa==4.9.1
ruff==0.12.7
safehttpx==0.1.6
selenium==4.34.2
semantic-version==2.10.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlalchemy==2.0.42
starlette==0.47.2
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.10.0
tomlkit==0.13.3
torch==2.8.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
triton==3.4.0
typer==0.16.0
typing-extensions==4.14.1
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wolframalpha==5.1.3
wsproto==1.2.0
xlrd==2.0.2
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.20.1
yt-dlp==2025.7.21
zstandard==0.23.0
